Split out r* libs into local rlib package
